% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_data.R
\name{all_tables}
\alias{all_tables}
\title{All Tables}
\usage{
all_tables()
}
\value{
data.frame, a list of all available tables
}
\description{
Get list of all available tables on \url{opendata.1212.mn}.
}
\examples{
all.tables <- all_tables()
print(all.tables$tbl_nm)
print(all.tables$tbl_id)
}
\seealso{
\link{statistical_data}
}
