% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sector_info.R
\name{get_sector_info}
\alias{get_sector_info}
\title{Get Sector}
\usage{
get_sector_info()
}
\value{
data frame, Sector information. It has following structures:
\enumerate{
 \item{rownum -}{ Row number}
 \item{list_id -}{ Sector identification number}
 \item{up_list_id -}{ Sub sector identification number}
 \item{list_nm -}{ Sector name /Mongolia/}
 \item{list_eng_nm -}{ Sector name /English/}
 \item{isExist -}{ Exist sub sector}
}
}
\description{
Get detailed information about all 34 sectors on \url{opendata.1212.mn}.
}
\examples{
sector_info <- get_sector_info()
print(sector_info)
}
\references{
\url{http://opendata.1212.mn/en/doc/Api/GET-api-Sector}
}
\seealso{
\link{all_tables}, \link{get_table}, \link{get_table_info}, \link{get_subsector_info}
}
