% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_info.R
\name{get_table_info}
\alias{get_table_info}
\title{Table Information and Classification}
\usage{
get_table_info(tbl_id)
}
\arguments{
\item{tbl_id}{character string, Table identification number}
}
\value{
list, Table information and classification. It has following structures:
\itemize{
 \item{tbl_id -}{ Table identification number}
 \item{unit_id -}{ Unit Identification number}
 \item{unit_nm -}{ Unit Name /Mongolia/}
 \item{unit_eng_nm -}{ Unit Name /English/}
 \item{obj -}{ Table classification:
  \itemize{
   \item{obj_var_id -}{ Variable identification}
   \item{var_ord_sn -}{ Variable identification number}
   \item{field -}{ Field name}
   \item{scr_mn -}{ Variable name /Mongolia/}
   \item{scr_eng -}{ Variable identification}
   \item{itm -}{ Variable classification and code:
    \itemize{
     \item{itm_id -}{ Classification Number}
     \item{up_itm_id -}{ Sub classification}
     \item{scr_mn -}{ Classification name /Mongolia/}
     \item{scr_eng -}{ Classification name /English/}
    }
   }
  }
 }
}
}
\description{
Get detailed information about a table on \url{opendata.1212.mn}.
}
\examples{
table_info <- get_table_info("DT_NSO_2400_015V2")
table_info$unit_nm
table_info$obj[[1]]$itm[[1]]$scr_mn
}
\references{
\url{http://opendata.1212.mn/en/doc/Api/GET-api-Itms-id}
}
\seealso{
\link{all_tables}, \link{get_table}, \link{get_sector_info}
}
