% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSO1212-package.r
\docType{package}
\name{NSO1212}
\alias{NSO1212}
\alias{NSO1212-package}
\title{NSO's Open Data API Handler}
\description{
Package providing helper functions for the National Statistical Office of Mongolia's open data API (\url{opendata.1212.mn}) version 2.
}
\author{
MAKHGAL Ganbold
}
