% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_data.R
\name{statistical_data}
\alias{statistical_data}
\title{Get statistical data}
\usage{
statistical_data(tbl_id, PERIOD = NULL, CODE = NULL, CODE1 = NULL,
  CODE2 = NULL)
}
\arguments{
\item{tbl_id}{character string, Table identification number}

\item{PERIOD}{charcter vector, Time}

\item{CODE}{charcter vector, Classification code (age, gender etc)}

\item{CODE1}{charcter vector, Classification code (age, gender etc)}

\item{CODE2}{charcter vector, Classification code (age, gender etc)}
}
\value{
data frame, Statistical data
}
\description{
Get statistical data from \url{opendata.1212.mn}.
}
\examples{
statistical.data <- statistical_data(
  tbl_id = "DT_NSO_2400_015V2",
  PERIOD = c("201711", "201712", "201801"),
  CODE = c("10", "11"),
  CODE1 = "11"
)
print(statistical.data)
}
\seealso{
\link{all_tables}
}
