% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sector_info.R
\name{get_sector_info}
\alias{get_sector_info}
\title{Detailed Information about All Sectors on the API opendata.1212.mn}
\usage{
get_sector_info()
}
\value{
data frame, Sector information. It has following structure:
\describe{
 \item{rownum}{Row number}
 \item{list_id}{Sector identification number}
 \item{up_list_id}{Sub sector identification number}
 \item{list_nm}{Sector name in Mongolian}
 \item{list_eng_nm}{Sector name in English}
 \item{isExist}{Whether or exist sub-sectors}
}
}
\description{
Detailed information about all 34 sectors, which are major classification of data, on the API \url{opendata.1212.mn} is supported by National Statistical Office of Mongolia (NSO)
}
\examples{
sector_info <- get_sector_info()
print(sector_info)
}
\references{
\url{http://opendata.1212.mn/en/doc/Api/GET-api-Sector}
}
\seealso{
\link{all_tables}, \link{get_table}, \link{get_table_info}, \link{get_subsector_info}
}
