% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSO1212-package.r
\docType{package}
\name{NSO1212}
\alias{NSO1212-package}
\alias{NSO1212}
\title{National Statistical Office of Mongolia's Open Data API Handler}
\description{
The National Statistical Office of Mongolia (NSO) is the national statistical service and an organization of the Mongolian government. NSO provides open access to official data via its open-data API. The package NSO1212 has functions for accessing the API service. The functions are compatible with API v2.0 and get data sets or their detailed information from the API.
}
\references{
\url{http://opendata.1212.mn/en/doc}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/makhgal-ganbold/NSO1212}
  \item Report bugs at \url{https://github.com/makhgal-ganbold/NSO1212/issues}
}

}
\author{
Makhgal Ganbold, National University of Mongolia
}
